backend/src/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.5
qdrant-client==1.7.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
cohere==5.5.3